requirements.txt
pymupdf
tiktoken
sentence-transformers
llama-cpp-python
flask-markdown

numpy 
matplotlib
tqdm
flask
